% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{AverageSampleModuleZscore}
\alias{AverageSampleModuleZscore}
\title{AverageSampleModuleZscore apply the function  calc_avg_module_zscore to a pseudobulklist.}
\usage{
AverageSampleModuleZscore(
  average.metacell.list,
  module.list,
  use.module.subset = TRUE,
  modules.subset.by.celltype = modules.test
)
}
\arguments{
\item{average.metacell.list}{poorly named argument - the object created by PseudobulkList either an average or summed pseudobulk data}

\item{module.list}{list of modules as named list each element is a vector of gene names}

\item{use.module.subset}{TRUE or FALSE - calc a different set of modules for each celltype use with modules.subset.by.celltype}

\item{modules.subset.by.celltype}{modules.subset.by.celltype is a list of modules to test with length = celltypes.vector and n modules = unique to subset}
}
\value{
returns a dataframe of module scores for each celltype
}
\description{
AverageSampleModuleZscore apply the function  calc_avg_module_zscore to a pseudobulklist.
}
\examples{
\dontrun{
AverageSampleModuleZscore = function(
 average.metacell.list = av,
 module.list = btm ,
 use.module.subset = FALSE,
 )

}
}
