% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_zscore.r
\name{calc_avg_module_zscore}
\alias{calc_avg_module_zscore}
\title{calc_avg_module_zscore - calculate average module z score of list of modules on a PseudobulkList
This is equivalent to the average z score method used in in Kotliarov et. al. Nature Med 2020
zscore is calculated across both genes and samples
it is adopted below to run on 'pseudobulk lists' (average "averagemetacell.list" or pseudobulk list
created by PseudobulkList) this small wrapper is called by the AverageSampleModuleZscore.
calculate signature score for each cell type, BTM, Subject
function input = named list of modules, dataframe with subject as rows genes as columns}
\usage{
calc_avg_module_zscore(module.list, average.data.frame)
}
\arguments{
\item{module.list}{list of modules}

\item{average.data.frame}{- this is created in AverageSampleModuleZscore}
}
\value{
see AverageSampleModuleZscore
}
\description{
calc_avg_module_zscore - calculate average module z score of list of modules on a PseudobulkList
This is equivalent to the average z score method used in in Kotliarov et. al. Nature Med 2020
zscore is calculated across both genes and samples
it is adopted below to run on 'pseudobulk lists' (average "averagemetacell.list" or pseudobulk list
created by PseudobulkList) this small wrapper is called by the AverageSampleModuleZscore.
calculate signature score for each cell type, BTM, Subject
function input = named list of modules, dataframe with subject as rows genes as columns
}
\examples{
\dontrun{
results = calc_avg_module_zscore(module.list = btm, average.data.frame = av_df)
}
# Average Module sample Z score
}
