% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{CombineResults}
\alias{CombineResults}
\title{CombineResults - For all cell types merge the GeneSetEnrichment leading edge genes coefficieint and p value from limma}
\usage{
CombineResults(gsealist, contrastlist, gseafdr, genefdr)
}
\arguments{
\item{genefdr}{}
}
\value{

}
\description{
CombineResults - For all cell types merge the GeneSetEnrichment leading edge genes coefficieint and p value from limma
}
\examples{
test = CombineResults(gsealist = testgsea, contrastlist = testmod, gseafdr = 0.05,genefdr = 0.1)
}
