% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{dreamMixedModel}
\alias{dreamMixedModel}
\title{dreamMixedModel - run dream mixed model}
\usage{
dreamMixedModel(
  dge_lists,
  apriori_contrasts = FALSE,
  sample_column,
  contrast_matrix = NULL,
  design_matrix,
  fixed_effects,
  cell_metadata,
  lme4_formula = "~ 0 + cohort_timepoint + (1|sampleid)",
  plotsavepath,
  ncores = 4
)
}
\arguments{
\item{dge_lists}{list of dgelists created with NormalizePseudobulk}

\item{apriori_contrasts}{one of TRUE or FALSE, whether to fit a priori contrasts}

\item{sample_column}{quoted character e.g. "sample" the subject level sample variable should have multiple timepoints subjectID_timepoint i.e. s1_0, s1_1}

\item{contrast_matrix}{contrast matrix created with make.contrasts}

\item{design_matrix}{design matrix created with BulkDesignMatrix}

\item{fixed_effects}{a vector of covariates that are columns in metadata e.g. a vector of  'age' 'gender'}

\item{cell_metadata}{metadata cells-rows variables-columns i.e. ColData or seuratmeta.data}

\item{lme4_formula}{symbolic model formula the default is '~ 0 + cohort_timepoint + (1|sampleid)'}

\item{plotsavepath}{a path to save created plot of contrasts}

\item{ncores}{number of cores for doParallel}
}
\value{
list of model fits indexed by celltype
}
\description{
dreamMixedModel - run dream mixed model
}
\examples{
\dontrun{
c_mat = makeContrasts(
foldchange_difference = (group_timepoint1_1 - group_timepoint1_0) - (group_timepoint0_1 - group_timepoint0_0),
time1_foldchange = (group_timepoint1_1 + group_timepoint0_1) / 2  - (group_timepoint1_0 + group_timepoint0_0) / 2,
baseline_groups = (group_timepoint1_0 - group_timepoint0_0),
levels = colnames(designmat)
)
# fit mixed model for the multi timepoint contrasts
fit = scglmmr::dreamMixedModel(dge_lists = dge,
apriori_contrasts = TRUE,
sample_column = 'sample',
cell_metadata = meta,
contrast_matrix = c_mat,
design_matrix = designmat,
lme4_formula =  '~ 0 + age + gender + cohort_timepoint + (1|sampleid)',
fixed_effects = c('age', 'gender', 'cohort_timepoint'),
plotsavepath = figpath,
ncores = 4)
}
# run dream mixed model
}
