% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream.wrapper.r
\name{FitDream}
\alias{FitDream}
\title{FitDream - run mixed effects model on aggregated (summed) data using the method 'dream'
by Hoffman et. al. Bioinformatics (2021) doi.org/10.1093/bioinformatics/btaa687. Fits
mixed model using lme4 with REML and voom weights.}
\usage{
FitDream(
  pb.list,
  sample.metadata,
  lme4.formula,
  dream.contrast.matrix = NULL,
  returnvoom = FALSE,
  ncores = 4,
  ...
)
}
\arguments{
\item{sample.metadata}{metadata, for example, object returned by AggregateCellMetadata}

\item{lme4.formula}{symbolic model formula for model to be fit, for example,
'~ 0 + group.timepoint + age + sex + (1|SubjectID)'. covariates must be in sample.metadata}

\item{ncores}{number of cores for biocparallel}

\item{dge.lists}{list of DGEList objects indexed by cell types -- the object returned by `scglmmr::Normalize`}

\item{sample_column}{quoted character e.g. "sample" the subject level sample variable should have multiple timepoints subjectID_timepoint i.e. s1_0, s1_1}
}
\value{
list of model fits indexed by celltype
}
\description{
FitDream - run mixed effects model on aggregated (summed) data using the method 'dream'
by Hoffman et. al. Bioinformatics (2021) doi.org/10.1093/bioinformatics/btaa687. Fits
mixed model using lme4 with REML and voom weights.
}
\examples{
\dontrun{
# make contrast matrix
L2 = makeContrastsDream(
  formula = f1,
  data = metadata,
  contrasts = c(
    baseline = "Group.time1_0 - Group.time0_0",
    treatment_delta = "( Group.time1_1 - Group.time1_0 ) - ( Group.time0_1 - Group.time0_0 )",
    treatment = "( Group.time1_1 + Group.time0_1 ) / 2 - ( Group.time1_0 + Group.time0_0 ) / 2 "
  )
)
f1 = '0 + group.time + age + sex + (1|SubjectID)'

}

fits = FitDream(pb.list = pb, sample.metadata = metadata, lme4.formula = f1, dream.contrast.matrix = L2, ncores = 4)
}
