% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{GetLeadingEdgeFull}
\alias{GetLeadingEdgeFull}
\title{GetLeadingEdgeFull Get a tidy dataframe of ALL Leading Edge Genes from gene set enrichment for all cell types}
\usage{
GetLeadingEdgeFull(gsea.list, padj.filter = NULL, NES.filter, p.filter = NULL)
}
\arguments{
\item{gsea.list}{results returned by RunFgseaOnRankList}

\item{padj.filter}{filter reslts}

\item{NES.filter}{filter results}

\item{p.filter}{raw p value filter -- padj.filter must not be specified}
}
\value{
a list
}
\description{
GetLeadingEdgeFull Get a tidy dataframe of ALL Leading Edge Genes from gene set enrichment for all cell types
}
\examples{
\dontrun{
lefull = scglmmr::GetLeadingEdgeFull(gsea.list = gsea1, padj.filter = 0.1,NES.filter = -Inf)
}
}
