% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetLeadingEdgeGenes}
\alias{GetLeadingEdgeGenes}
\title{GetLeadingEdgeGenes get the leading edge genes from a single cell type / module combination}
\usage{
GetLeadingEdgeGenes(gsea.result.list, celltype.index, module.name)
}
\arguments{
\item{gsea.result.list}{results from RunFgseaOnRank}

\item{celltype.index}{celltype number of result list}

\item{module.name}{name of module}
}
\value{

}
\description{
GetLeadingEdgeGenes get the leading edge genes from a single cell type / module combination
}
