% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetRankResults}
\alias{GetRankResults}
\title{GetRankResults get list of gene ranks by t stat for fGSEA}
\usage{
GetRankResults(limma.fit.object.list, coefficient.number, contrast.name)
}
\arguments{
\item{limma.fit.object.list}{the results returned by RunLimmaVoom. Use GetRankResultsRaw for results returned by dreamMixedModel}

\item{contrast.name}{}
}
\value{

}
\description{
GetRankResults get list of gene ranks by t stat for fGSEA
}
