% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetRankResultsRaw}
\alias{GetRankResultsRaw}
\title{GetRankResults get list of gene ranks by t stat for fGSEA. for results returned by dreamMixedModel}
\usage{
GetRankResultsRaw(contrast.result.raw.list)
}
\arguments{
\item{contrast.result.raw.list}{results returned by dreamMixedModel}
}
\value{

}
\description{
GetRankResults get list of gene ranks by t stat for fGSEA. for results returned by dreamMixedModel
}
