% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_result_interaction.r
\name{GetTidySummary}
\alias{GetTidySummary}
\title{GetTidySummary - tidy data summary for a single cell type}
\usage{
GetTidySummary(av.exprs.list, celltype.index, genes.use)
}
\arguments{
\item{av.exprs.list}{- object returned by PseudobulkList (use average or first convert summed counts to cpm)}

\item{celltype.index}{- index of celltype to ret results see names of PseudobulkList object}

\item{genes.use}{- subet of genes to use}
}
\value{
a tidy dataframe
}
\description{
GetTidySummary - tidy data summary for a single cell type
}
\examples{
#' @examples
\dontrun{
gene_highlight =  c("IRF1","TNFRSF17","ABL1")
mono = GetTidyCohort(av.exprs.list = av, celltype.index = 7, genes.use = gene_highlight)
PlotGeneDistCohort(merged_av_data = mono,
                   save_name = "mono_highlight_512",
                   save_path = figpath,
                   title = paste0(names(av[7]), "genesub" ),
                   height = 3.8, width = 4.5,
                   nrow = 2)
 }
}
