% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GSEABarPlot}
\alias{GSEABarPlot}
\title{GSEABarPlot - plot gsea results for a single cell type}
\usage{
GSEABarPlot(
  rbind_gsea_result_dataframe,
  celltype_name,
  save_path,
  title,
  save_name,
  fill_color,
  width = 8.5,
  height = 7.2
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
GSEABarPlot - plot gsea results for a single cell type
}
