% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{LeadEdgeSampleHeatmap}
\alias{LeadEdgeSampleHeatmap}
\title{LeadEdgeSampleHeatmap make a heatmap of average expression for top or leading edge genes returned by LeadEdgeTidySampleExprs}
\usage{
LeadEdgeSampleHeatmap(
  tidy.exprs.list,
  modulename,
  celltype_plot,
  metadata,
  metadata_annotate,
  sample_column,
  returnmat = FALSE,
  plotwidth = 5,
  plotheight = 8,
  savepath,
  savename
)
}
\arguments{
\item{tidy.exprs.list}{the object returned by returned by TopGenesTidySampleExprs or LeadEdgeTidySampleExprs which is created from metadata adn the object returned by PseudobulkList (can be average or summed expression)}

\item{modulename}{The name of the module to plot}

\item{celltype_plot}{the name of the celltype to plot}

\item{metadata}{cells x metadata dataframe}

\item{metadata_annotate}{a vector of variables (columns) of sample level  metadata to annotate on the heatmap. Must be categorical for each sample e.g. "age" but not "nUMI"}

\item{sample_column}{the column in the metadata object corresponding to the sample labels, usually 'smaple'}

\item{returnmat}{instead of making an annotated heatmap just return the matrix of averge values per sample of the module subset}
}
\value{
a pheatmap object
}
\description{
LeadEdgeSampleHeatmap make a heatmap of average expression for top or leading edge genes returned by LeadEdgeTidySampleExprs
}
