% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{LeadingEdgeIndexed}
\alias{LeadingEdgeIndexed}
\title{LeadingEdgeIndexed - extract leading edge genes from FgseaList results, return a new embedded list of named modules, indexed by celltype}
\usage{
LeadingEdgeIndexed(gsea.result.list, padj.threshold = 0.05)
}
\arguments{
\item{gsea.result.list}{results from FgseaList or RunFgseaOnRankList}

\item{padj.threshold}{within each cell type return list of leading edge genes with padj less than this parameter value.}
}
\value{
embedded list - each list level 1 indexed by cell type, contains a new list level 2, the leading edge genes from the gsea results filtered by padj.
}
\description{
LeadingEdgeIndexed - extract leading edge genes from FgseaList results, return a new embedded list of named modules, indexed by celltype
}
\examples{
\dontrun{
t1hvl_rank = GetRankResultsRaw(limma.fit.object.list  = ebf,
coefficient.number = 1,
contrast.name = "time_1_highvslow")
gsea = FgseaList(rank.list.celltype = t1hvl_rank)
celltype.indexed.modules.leadingedge = LeadingEdgeIndexed(gsea, 0.05)
}
}
