% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{PlotFgsea}
\alias{PlotFgsea}
\title{PlotFgsea identical to GSEABubblePlot, returns plot for manual adjustment or saving and also clusters the map.}
\usage{
PlotFgsea(gsea_result_list, NES_filter = -Inf, padj_filter = 0.1)
}
\arguments{
\item{rbind_gsea_result_dataframe}{dataframe returned by RbindGseaResultList}
}
\value{
ggplot object
}
\description{
PlotFgsea identical to GSEABubblePlot, returns plot for manual adjustment or saving and also clusters the map.
}
\examples{
\dontrun{
scglmmr::GSEABubblePlot(d, save_path = figpath, save_name = "plot.pdf")
}
}
