% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{PlotFgseaList}
\alias{PlotFgseaList}
\title{PlotFgseaList identical to GSEABubblePlot, returns plot for manual adjustment or saving.}
\usage{
PlotFgseaList(gsea_result_dataframe)
}
\arguments{
\item{rbind_gsea_result_dataframe}{dataframe returned by RbindGseaResultList}
}
\value{
ggplot object
}
\description{
PlotFgseaList identical to GSEABubblePlot, returns plot for manual adjustment or saving.
}
\examples{
\dontrun{
scglmmr::GSEABubblePlot(d, save_path = figpath, save_name = "plot.pdf")
}
}
