% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{PlotGeneDistCohort}
\alias{PlotGeneDistCohort}
\title{Title}
\usage{
PlotGeneDistCohort(
  merged_av_data,
  save_path,
  save_name,
  title = NULL,
  nrow = 5,
  height = 8,
  width = 10,
  plot_subset = FALSE,
  genes_plot = NULL
)
}
\arguments{
\item{merged_av_data}{data returned for a single cell type by `GetTidySummary`}

\item{save_path}{file path to save results}

\item{save_name}{name of plot saved to `save_path`}

\item{title}{title of plot}

\item{nrow}{number of rows to facet genes plotted on}

\item{height}{ggsave param}

\item{width}{ggsave param}

\item{plot_subset}{whether to subset to some of the genes in `merged_av_data`}

\item{genes_plot}{the subset of genes to plot}
}
\value{
nothing
}
\description{
Title
}
