% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{PlotHypergeometric}
\alias{PlotHypergeometric}
\title{PlotHypergeometric - plot results returned by RunHypergeometricTest}
\usage{
PlotHypergeometric(
  hyperg_result,
  p.adjust.filter = 0.05,
  genenumber_filter = 0,
  savepath = figpath,
  savename,
  title,
  height = 10,
  width = 8
)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
PlotHypergeometric - plot results returned by RunHypergeometricTest
}
