% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{PseudobulkList}
\alias{PseudobulkList}
\title{PseudobulkList - make a pseudobuk summed or average dataset for each samplexcelltype}
\usage{
PseudobulkList(
  rawcounts,
  metadata,
  sample_col,
  celltype_col,
  avg_or_sum = "sum"
)
}
\arguments{
\item{rawcounts}{the raw count UMI data for assay to sum or average for each celtype and sample.}

\item{metadata}{dataframe of meta data for cells-rows variables-columns i.e. ColData or seuratmeta.data}

\item{sample_col}{quoted character e.g. "sample" the subject level sample variable - if multiple timepoints should be subjectID_timepoint i.e. s1_0, s1_1}

\item{celltype_col}{quoted character e.g. "celltype" - the celltypes / clusters for which to create bulk libraries}

\item{avg_or_sum}{whether to compute default 'sum' or 'average' or average library for each samplexcelltype}
}
\value{
a R list of standard R matrices indexed by celltype
}
\description{
PseudobulkList - make a pseudobuk summed or average dataset for each samplexcelltype
}
\examples{
# for each celltype aggregate raw counts for ech sample (subject timepoint combination)
}
