% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{RunFgseaOnRankList}
\alias{RunFgseaOnRankList}
\title{RunFgseaOnRankList - wrapper around fast gene set enrichment analysis with the fgsea R package https://bioconductor.org/packages/release/bioc/html/fgsea.html}
\usage{
RunFgseaOnRankList(
  rank.list.celltype,
  pathways,
  maxSize = 500,
  minSize = 9,
  nperm = 250000,
  positive.enrich.only = FALSE
)
}
\arguments{
\item{rank.list.celltype}{results returned by GetRankResultsRaw or GetRankResults}

\item{pathways}{modules / gene sets as a named list each a single vector of unique gene IDS}

\item{maxSize}{see fgsea package}

\item{minSize}{see fgsea package}

\item{nperm}{recommended to keep set at 25000 based on optomization and p value stabilization}

\item{positive.enrich.only}{include negative enrichments in results? TRUE/FALSE}
}
\value{
results from fgsea package indexed by celltype
}
\description{
RunFgseaOnRankList - wrapper around fast gene set enrichment analysis with the fgsea R package https://bioconductor.org/packages/release/bioc/html/fgsea.html
}
\examples{
\dontrun{
t1hvl_rank = GetRankResultsRaw(limma.fit.object.list  = ebf, coefficient.number = 1, contrast.name = "time_1_highvslow")
gsea = RunFgseaOnRankList(rank.list.celltype = t1hvl_rank)
}
# usage:
}
