% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{RunHypergeometricTest}
\alias{RunHypergeometricTest}
\title{RunHypergeometricTest - run a hypergeometric test on results returned by GetContrastResults or GetContrastResultsRaw}
\usage{
RunHypergeometricTest(
  result_list,
  TERM2GENE_dataframe,
  pval_threshold = 0.05,
  logFC_threshold = 0.5,
  usefdr_threshold = FALSE
)
}
\arguments{
\item{result_list}{results returned by by GetContrastResults or GetContrastResultsRaw}

\item{TERM2GENE_dataframe}{see clusterprofiler, these objects are automatically loaded in the scglmmr package one of term_df_btm, term_df_kegg, term_df_reactome etc.}

\item{pval_threshold}{p threshold for genes to consider in hypergeometric distribution}

\item{logFC_threshold}{logFC threshold  for genes to consider in hypergeometric distribution}

\item{usefdr_threshold}{use the FDR adjusted p values for ranking genes-this is a strict filter for single cell data, recommended to set FALSE}
}
\value{

}
\description{
RunHypergeometricTest - run a hypergeometric test on results returned by GetContrastResults or GetContrastResultsRaw
}
\examples{
hyp_hvl = RunHypergeometricTest(result_list = d1red, TERM2GENE_dataframe = term_df_btm, pval_threshold = 0.1, logFC_threshold = 0, usefdr_threshold = FALSE)
}
