% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{RunVoomLimma}
\alias{RunVoomLimma}
\title{RunVoomLimma}
\usage{
RunVoomLimma(dgelists, design_matrix, do_contrast_fit, my_contrast_matrix)
}
\arguments{
\item{dgelists}{a ist of DGElist created in NormalizePseudobulk}

\item{design_matrix}{design matrix created with BulkDesignMatrix}

\item{do_contrast_fit}{whether to fit custom a priori contrasts}

\item{my_contrast_matrix}{custom a priori contrasts created with make.contrasts - see limma or edgeR manual}
}
\value{
a list of linear model fits for each celltype
}
\description{
RunVoomLimma
}
\examples{
# run limma using voom observational weights for non mixed effects models using emperical bayes
}
