% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{RunVoomLimma_covar}
\alias{RunVoomLimma_covar}
\title{Title}
\usage{
RunVoomLimma_covar(
  dgelists,
  design_matrix,
  co_variable_genes = NULL,
  grouptable,
  do_contrast_fit,
  my_contrast_matrix,
  my_model_metadata,
  celltypes.vector = NULL
)
}
\arguments{
\item{dgelists}{DGElists created with PseudobulkList}

\item{design_matrix}{design matrix created with BulkDesignMatrix}

\item{co_variable_genes}{co variable genes}

\item{grouptable}{group table}

\item{do_contrast_fit}{do contrasts}

\item{my_contrast_matrix}{contrast matrix}

\item{my_model_metadata}{model metadata}

\item{celltypes.vector}{vector of celltypes assign as names of dgeList}
}
\value{

}
\description{
Title
}
