% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupContrastGLMMsinglecell.r
\name{SCGroupContrastGLMM}
\alias{SCGroupContrastGLMM}
\title{SCGroupContrastGLMM - within each cell type contrast difference in fold change between groups, baseline difference, and fold change across groups of module scores.}
\usage{
SCGroupContrastGLMM(
  module_data_frame,
  celltype_column = "celltype",
  metadata,
  fixed_effects = NULL,
  lmer_formula = NULL,
  plotdatqc = TRUE,
  figpath
)
}
\arguments{
\item{module_data_frame}{data for each cell to model -- designed to be scores for modules (columns) returned by scglmmr::WeightedModuleScore}

\item{celltype_column}{the column in metadata with the cluster / celltype designation for each cell}

\item{lme4metadata}{metadata for model fit}

\item{f1}{model furmula}

\item{contrast_list}{list of linear model contrasts.}

\item{plot_savepath}{path to save results}
}
\value{

}
\description{
SCGroupContrastGLMM - within each cell type contrast difference in fold change between groups, baseline difference, and fold change across groups of module scores.
}
