% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMixedPoisson.r
\name{SCMixedPoisson}
\alias{SCMixedPoisson}
\title{SCMixedPoisson}
\usage{
SCMixedPoisson(
  gene_data,
  lme4metadata,
  model_formula = "gene ~ offset(log(nUMI)) + timepoint + (1|subjectid)",
  celltype_genes_test = NULL
)
}
\arguments{
\item{gene_data}{matrix of raw single cell UMI counts genes as COLUMNS and cells as ROWS; i.e. a transpose t() of Bioconductor or Seurat raw.data slots if genes_test is specified, columns must contain all genes from genes_test i.e. umi_matrix = t(seurat_raw_counts[unique(unlist((genes_test)), ])}

\item{lme4metadata}{metadata for each cell and cell barcodes as}

\item{model_formula}{defaults to 'gene ~ offset(log(nUMI)) + timepoint + (1|subjectid)' the variable 'timepoint' can be altered to be any treatment or perturbation effect being tested. Other variables should be kept the same.}

\item{celltype_genes_test}{A R list indexed by cell type of the subset of genes to test for each cell type.}

\item{reduced_model_formula}{identical formula as `full_model_formula` without the treatment or perturbation variable term. Defaults to 'gene ~ offset(log(nUMI)) + (1|subjectid)'}
}
\value{
a results matrix for all genes across all cell types tested.
}
\description{
SCMixedPoisson
}
\examples{
results = SCMixedPoisson(gene_data = gene_data,
  lme4metadata = meta_data,
  model_formula = 'gene ~ offset(log(nUMI)) + timepoint + (1|sampleid)',
  celltype_genes_test = celltype_indexed_gene_vector
)



}
