% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{TopGenesTidySampleExprs}
\alias{TopGenesTidySampleExprs}
\title{TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast}
\usage{
TopGenesTidySampleExprs(
  av.exprs.list,
  result.list,
  P.Value.filter,
  logFC.filter,
  top_n_genes = NULL
)
}
\arguments{
\item{av.exprs.list}{object returned by `PseudobulkList` summed or average counts}

\item{result.list}{object returned by object returned by `GetContrastResultsRaw()` or `GetContrastResults()`}

\item{P.Value.filter}{filter results}

\item{logFC.filter}{filter results}

\item{top_n_genes}{instead of stat thresholds above, get expression of n genes specified by this param, subsets each celltype by the top n genes  ranked by t statistic.}
}
\value{
a list of tidy dataframes by celltype
}
\description{
TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast
}
