% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_module_score_functions.r
\name{WilcoxWithinCluster}
\alias{WilcoxWithinCluster}
\title{WilcoxWithinCluster}
\usage{
WilcoxWithinCluster(module_data_frame, lme4metadata, plot_savepath = NULL)
}
\arguments{
\item{plot_savepath}{}
}
\value{

}
\description{
WilcoxWithinCluster
}
