% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{calc_avg_module_zscore}
\alias{calc_avg_module_zscore}
\title{calc_avg_module_zscore calculate average module z score of list of modules on a PseudobulkList}
\usage{
calc_avg_module_zscore(module.list, average.data.frame)
}
\arguments{
\item{module.list}{list of modules}

\item{average.data.frame}{- this is created in AverageSampleModuleZscore}
}
\value{
see AverageSampleModuleZscore
}
\description{
calc_avg_module_zscore calculate average module z score of list of modules on a PseudobulkList
}
\examples{
# Average Module sample Z score
}
