% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{CombineResults}
\alias{CombineResults}
\title{CombineResults - For all cell types merge the gsea leading edge genes with their contrast model coefficieint and p value from limma / dream}
\usage{
CombineResults(gsealist, contrastlist, gseafdr, gseap, genefdr = Inf)
}
\arguments{
\item{gsealist}{list of results returned by `RunFgseaOnRankList()`}

\item{contrastlist}{list of results returned by  `scglmmr::ExtractResult()` or from older versions of scglmmr, the equivalent results from: `GetContrastResults()` , `GetContrastResultsRaw()`}

\item{gseafdr}{the adjusted p value threshold to filter out results (gsea)}

\item{gseap}{the p value threshold to filter out results (gsea); gseafdr must not be specified to use this option}

\item{genefdr}{the adjusted p value threshold to filter out individual genes - by default all the leading edge genes are returned (recommended)}
}
\value{
a tidy dataframe
}
\description{
CombineResults - For all cell types merge the gsea leading edge genes with their contrast model coefficieint and p value from limma / dream
}
\examples{
\dontrun{
combined_results = CombineResults(gsealist = testgsea, contrastlist = testmod, gseafdr = 0.05,genefdr = 0.2)
}
}
