% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetLeadingEdgeFull}
\alias{GetLeadingEdgeFull}
\title{GetLeadingEdgeFull Get a tidy dataframe of ALL Leading Edge Genes from gene set enrichment for all cell types}
\usage{
GetLeadingEdgeFull(gsea.list, padj.filter, NES.filter)
}
\arguments{
\item{gsea.list}{results returned by RunFgseaOnRankList}

\item{padj.filter}{filter reslts}

\item{NES.filter}{filter results}
}
\value{

}
\description{
GetLeadingEdgeFull Get a tidy dataframe of ALL Leading Edge Genes from gene set enrichment for all cell types
}
\examples{
y = GetLeadingEdgeFull(gsea, padj.filter = 0.1, NES.filter = -Inf)
}
