% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{GetLeadingEdgeGenes}
\alias{GetLeadingEdgeGenes}
\title{GetLeadingEdgeGenes get the leading edge genes from a single cell type / module combination}
\usage{
GetLeadingEdgeGenes(gsea.result.list, celltype.index, module.name)
}
\arguments{
\item{gsea.result.list}{results from RunFgseaOnRank}

\item{celltype.index}{celltype number of result list}

\item{module.name}{name of module}
}
\value{
a vector of genes
}
\description{
GetLeadingEdgeGenes get the leading edge genes from a single cell type / module combination
}
\examples{
\dontrun{
le_mono = GetLeadingEdgeGenes(gsea.result.list = gsea1, celltype.index = 4, module.name = 'my_modulename_from_gsearesults')
}
}
