% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetRankResultsRaw}
\alias{GetRankResultsRaw}
\title{GetRankResultsRaw get list of gene ranks by t stat for fGSEA. for results returned by dreamMixedModel}
\usage{
GetRankResultsRaw(contrast.result.raw.list)
}
\arguments{
\item{contrast.result.raw.list}{results returned by dreamMixedModel}
}
\value{
list of gene ranks by t stat -- use as argument to `RunFgseaOnRankList`
}
\description{
GetRankResultsRaw get list of gene ranks by t stat for fGSEA. for results returned by dreamMixedModel
}
\examples{
\dontrun{
fit_res = scglmmr::GetContrastResultsRaw(limma.fit.object.list = fit,
                                         coefficient.number = 1,
                                         contrast.name = "foldchangedifference")
fit_rank = scglmmr::GetRankResultsRaw(contrast.result.raw.list = fit_res)
}
}
