% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{GetTidySummary}
\alias{GetTidySummary}
\title{GetTidySummary - tidy data summary for a single cell type}
\usage{
GetTidySummary(av.exprs.list, celltype.index, genes.use)
}
\arguments{
\item{av.exprs.list}{- object returned by PseudobulkList}

\item{celltype.index}{- index of celltype to ret results see names of PseudobulkList object}

\item{genes.use}{- subet of genes to use}
}
\value{

}
\description{
GetTidySummary - tidy data summary for a single cell type
}
