% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupContrastGLMMtidy.r
\name{GroupContrastGLMMtidy}
\alias{GroupContrastGLMMtidy}
\title{GroupContrastGLMMtidy - fit a random intercept model of the difference in fold change between groups and fold change across both groups and baseline controlling for any number of covariates.}
\usage{
GroupContrastGLMMtidy(
  tidydf,
  response_variable,
  response_value,
  fixed_effects = NULL,
  lmer_formula = NULL,
  figpath
)
}
\arguments{
\item{tidydf}{tidy data must contain column 'subjectid' and any names covariates (specified with )}

\item{response_variable}{the column containing the variables to be modeled in tidy (long) format -- e.g. 'gene' with the gene's values in the column in response_value}

\item{response_value}{the value for the measurement response_variable -- should probably be in log space.}

\item{fixed_effects}{the fixed effects to specify in this case as covariates e.g. c('gender', 'ethnicity'), these are automatically combined into formula with your_fixed_effects + 0 + group_id + (1|subjectid)}

\item{lmer_formula}{lmer model formula, if not specified, creates a donor random intercept model of 0 + group_id + (1|subjectid)}

\item{figpath}{path to save figures e.g.  file.path()}
}
\value{

}
\description{
GroupContrastGLMMtidy - fit a random intercept model of the difference in fold change between groups and fold change across both groups and baseline controlling for any number of covariates.
}
\examples{
\donttest{
  tidydf = d2 \%>\%
  dplyr::rename('subjectid'= subject) \%>\%
  mutate(group_id = factor(group_id, levels = c(0_0, 0_1, 1_0, 1_1)))

  response_variable = "logprot"
  fixed_effects = c("Age", "Race", "Gender")
  mmres = GroupContrastGLMMtidy(tidydf = tidydf, response_variable = 'protein', response_value = 'log10protvalue', fixed_effects = c('gender', 'ethnicity'), lmer_formula = NULL, subject_colum 'subjectid', figpath = '~' )
  }

}
