% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.r
\name{GSEABarPlot}
\alias{GSEABarPlot}
\title{GSEABarPlot - plot gsea results for a single cell type}
\usage{
GSEABarPlot(
  rbind_gsea_result_dataframe,
  celltype_name,
  fill_color,
  text.size = 8
)
}
\arguments{
\item{rbind_gsea_result_dataframe}{result returned from RbindGseaResultList}

\item{celltype_name}{name of celltype to be plotted}

\item{fill_color}{color of bar}

\item{text.size}{size of axis labels}
}
\value{
ggplot object
}
\description{
GSEABarPlot - plot gsea results for a single cell type
}
\examples{
\dontrun{
p = scglmmr::GSEABarPlot(d, celltype_name = 'celltype1', fill_color = 'dodgerblue')
}
}
