% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{PlotHypergeometric}
\alias{PlotHypergeometric}
\title{PlotHypergeometric - plot results returned by RunHypergeometricTest}
\usage{
PlotHypergeometric(
  hyperg_result,
  p.adjust.filter = 0.05,
  genenumber_filter = 0,
  savepath = figpath,
  savename,
  title,
  height = 10,
  width = 8
)
}
\arguments{
\item{hyperg_result}{result returned by `RunHypergeometricTest`}

\item{p.adjust.filter}{filter results}

\item{genenumber_filter}{number of genes within an enrichment minimum.}

\item{savepath}{save oath}

\item{savename}{name of object saved to save_path}

\item{title}{title of plot}

\item{height}{ggsave param}

\item{width}{ggsave param}
}
\value{
nothing
}
\description{
PlotHypergeometric - plot results returned by RunHypergeometricTest
}
