% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{PlotHypergeometric}
\alias{PlotHypergeometric}
\title{PlotHypergeometric - plot results returned by RunHypergeometricTest}
\usage{
PlotHypergeometric(
  hyperg_result,
  p.adjust.filter = 0.05,
  genenumber_filter = 0,
  savepath = figpath,
  savename,
  title,
  height = 10,
  width = 8
)
}
\arguments{
\item{hyperg_result}{result returned by `RunHypergeometricTest`}

\item{p.adjust.filter}{filter results}

\item{genenumber_filter}{number of genes within an enrichment minimum.}

\item{savepath}{save oath}

\item{savename}{name of object saved to save_path}

\item{title}{title of plot}

\item{height}{ggsave param}

\item{width}{ggsave param}
}
\value{
nothing
}
\description{
PlotHypergeometric - plot results returned by RunHypergeometricTest
}
\examples{
\dontrun{
load(termdf) # this term2gene dataframe is included in the package see clusterProfiler
hyp = scglmmr::RunHypergeometricTest(result_list = fit_res,
                                     TERM2GENE_dataframe = termdf,
                                     pval_threshold = 0.1,
                                     logFC_threshold = 0,
                                     usefdr_threshold = FALSE)
# plot results
scglmmr::PlotHypergeometric(hyperg_result = hyp,
                            p.adjust.filter = 0.1,
                            genenumber_filter = 2,
                            savepath = figpath,
                            savename = "name",
                            title = "title")
 }
}
