% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{RbindGseaResultList}
\alias{RbindGseaResultList}
\title{RbindGseaResultList - prepare gsea result list for visualization funcitons GseaBubblePlot or GseaBarPlot}
\usage{
RbindGseaResultList(gsea_result_list, NES_filter = -Inf, padj_filter = 0.1)
}
\arguments{
\item{padj_filter}{}
}
\value{

}
\description{
RbindGseaResultList - prepare gsea result list for visualization funcitons GseaBubblePlot or GseaBarPlot
}
