% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_de.R
\name{SubjectCelltypeTable}
\alias{SubjectCelltypeTable}
\title{SubjectCelltypeTable - make a QC contingency table of cells by samples beore making pseudobulk data.}
\usage{
SubjectCelltypeTable(metadata, celltype_column, sample_column)
}
\arguments{
\item{metadata}{metadata cells-rows variables-columns i.e. ColData or seuratmeta.data}

\item{celltype_column}{quoted character e.g. "celltype" - the celltypes / clusters for which to create bulk libraries}

\item{sample_column}{quoted character e.g. "sample" the subject level sample variable - if multiple timepoints should be subjectID_timepoint i.e. s1_0, s1_1}
}
\value{
a list with index 1 = contingency table of cells by sample, 2 = low representation celltypes < 10 cels, 3 celltypes to remove with 0 cells for some samples.
}
\description{
SubjectCelltypeTable - make a QC contingency table of cells by samples beore making pseudobulk data.
}
