% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis_functions.r
\name{TopGenesTidySampleExprs}
\alias{TopGenesTidySampleExprs}
\title{TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast}
\usage{
TopGenesTidySampleExprs(
  av.exprs.list,
  result.list,
  P.Value.filter,
  logFC.filter,
  top_n_genes = NULL
)
}
\arguments{
\item{top_n_genes}{}
}
\value{

}
\description{
TopGenesTidySampleExprs convert a PseudobulkList into a tidy dataframe for each sample across cell types of the top differentially expressed genes for a contrast
}
\examples{
topav = GetTopAvg(av.exprs.list = av,result.list = d1time, P.Value.filter = 0.01,logFC.filter = 1.2)
}
