% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCGroupContrast_modulefunctions.r
\name{WilcoxWithinCluster}
\alias{WilcoxWithinCluster}
\title{WilcoxWithinCluster}
\usage{
WilcoxWithinCluster(module_data_frame, lme4metadata, plot_savepath = NULL)
}
\arguments{
\item{plot_savepath}{}
}
\value{

}
\description{
WilcoxWithinCluster
}
\examples{
\dontrun{
weighted_score = WeightedCellModuleScore(seurat_object = s, module_list = sig,
                                         threshold = 0.1, cellwise_scaling = FALSE,
                                         return_weighted = TRUE,
                                         Seurat_version = "3")

# set module dataframe
module_df = cbind(celltype_joint = as.character(s@meta.data$seurat_clusters),
                  barcode_check = s@meta.data$barcode_full,
                  weighted_score) \%>\%
  droplevels()

# format lme4 metadata (note here just wilcox so group IDS do not have to be lme4 factors
meta_data = s@meta.data \%>\%
  droplevels() \%>\%
  select(celltype_joint = seurat_clusters,  barcode_check = barcode_full, sample, group_id = IRAE) \%>\%
  filter(group_id \%in\% c("poor_outcome", "good_outcome")) \%>\%
  mutate(group_id = factor(group_id, levels = c("poor_outcome", "good_outcome")))

# Fit wilcox test and do bonferonni correction
plot_savepath2 = file.path(figpath, "module_distribution/"); dir.create(plot_savepath2)
m2 = module_df \%>\% filter(barcode_check \%in\% meta_data$barcode_check )
cvnc = WilcoxWithinCluster(module_data_frame = m2, lme4metadata = meta_data, plot_savepath = plot_savepath2)
cvnc$padj = p.adjust(p = cvnc$p.value.t0.wilcox, method = "BH")
}
}
